fix: sidebar is now scrolled to the right position when unhiding
this was caused by the outer task updating both the sidebar and detail views to the value of selection
 simultaneously and regardless of whether the sidebar is hidden or not.
Changes to the scrollPosition binding require the values to actually differ in order to take effect.
So if sidebarScrollPosition is assigned 5 while the sidebar is gone, and then assigned 5 again when it
becomes visible, the binding won't update, as the value is the same.
Separating the assignments to different tasks attached to their respective views fixed that, as the
sidebar would stop updating when it's gone, and update again as soon as it shows.

Feels correct 🥰